Separated the Information Window into its own component. Added the infrastructure to render a preview of the currently selected object (its material included, that's the idea...). The goal is to have a material editor that will alter the object in the main scene.

+++ raytracer/main/ray.h
@@ -22,7 +22,7 @@ public:
         return m_direction;
     }
 
-    inline Point position(const double t) const {
+    inline Point position(const float t) const {
         return m_origin + m_direction * t;
     }
